separated influence function syntax from estimation semantics
